Move Query Assertions to new utility
