+++ graphql-client-processor/src/test/groovy/co/uk/jacobmountain/IntegrationSpec.groovy
@@ -16,6 +16,7 @@ class IntegrationSpec extends Specification {
         character.name = name
         character
     }
+
     static def newDroid(int id, String name) {
         def character = new Droid()
         character.id = id
@@ -25,7 +26,7 @@ class IntegrationSpec extends Specification {
 
     def "Can query for data with no args"() {
         given:
-        fetcher.query("query Hero { hero { id name friends { id name ... on Human { id name totalCredits __typename } ... on Droid { id name primaryFunction __typename } __typename } ... on Human { id name friends { id name ... on Human { id name totalCredits __typename } ... on Droid { id name primaryFunction __typename } __typename } totalCredits __typename } ... on Droid { id name friends { id name ... on Human { id name totalCredits __typename } ... on Droid { id name primaryFunction __typename } __typename } primaryFunction __typename } __typename } }", null) >> Response.builder()
+        fetcher.query(_, null) >> Response.builder()
                 .data(new Query() {
                     @Override
                     Character getHero() {
@@ -48,7 +49,7 @@ class IntegrationSpec extends Specification {
         def args = new GetDroidById()
         args.id = id
         def droid = newDroid(id, name)
-        fetcher.query("query Droid(\$id: ID!) { droid(id: \$id) { id name friends { id name ... on Human { id name totalCredits __typename } ... on Droid { id name primaryFunction __typename } __typename } primaryFunction __typename } }", args) >> Response.builder()
+        fetcher.query(_, args) >> Response.builder()
                 .data(new Query() {
                     @Override
                     Droid getDroid() {
